playsite: Factor out evy syntax highlighting for reuse in docs
Factor out evy syntax highlighting for reuse in docs. Initially we will not
use an editor in the docs so all we need to do is replace the <pre><code>
codeblocks with the highlighted content.
