+++ frontend/play/css/syntax.css
@@ -1,0 +1,1 @@
+../../css/syntax.css
