+++ frontend/play/index.html
@@ -12,13 +12,15 @@
     <link rel="stylesheet" href="css/icons.css" type="text/css" />
     <link rel="stylesheet" href="css/header.css" type="text/css" />
     <link rel="stylesheet" href="css/dialog.css" type="text/css" />
+    <link rel="stylesheet" href="css/syntax.css" type="text/css" />
     <link rel="stylesheet" href="css/index.css" type="text/css" />
     <link rel="stylesheet" href="css/fonts.css" type="text/css" />
     <script type="importmap">
       {
         "imports": {
           "./module/confetti.js": "./module/confetti.js",
-          "./module/editor.js": "./module/editor.js"
+          "./module/editor.js": "./module/editor.js",
+          "./module/highlight.js": "./module/highlight.js"
         }
       }
     </script>
@@ -58,7 +60,7 @@
     <div class="max-width-wrapper">
       <main class="main">
         <div class="editor-wrap noscrollbar">
-          <div class="editor" style="padding-left: calc(2ch + 1.5rem)">
+          <div class="editor language-evy" style="padding-left: calc(2ch + 1.5rem)">
             <!-- These editor sample contents are replaced by JS, once evy toolchain and editor are ready. -->
             <pre class="lines">
 <span class="num">  1</span>
