+++ frontend/play/module/highlight.js
@@ -1,0 +1,1 @@
+../../module/highlight.js
